Implements image upload and preview in plan creation
Adds functionality to upload and preview images when creating a new plan.

This change introduces an image upload area in the "add new plan" modal,
allowing users to select multiple images for their plan. Selected images
are displayed as thumbnails with a remove option. It also adds a user
notification when selecting non-image files.