Adds customer creation modal to dashboard
Enables adding new customers directly from the dashboard.

This commit introduces a modal for adding new customers, accessible
via a button on the dashboard.  It also updates the modal close
functionality to dynamically identify and close the appropriate modal.
Finally, it adds a click event to trigger the hidden file input.